+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 ### Bugs fixed
 
 ### Enhancements
-- GP2-3123-remove-redundant-code
+- GP2-3123 - remove redundant code
+- GP2-2857 - refactor countries import
 
 ## [2.00.0](https://github.com/uktrade/directory-api/releases/tag/2.00.0)
 [Full Changelog](https://github.com/uktrade/directory-api/compare/1.14.0...2.00.0)